backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
python-jose==3.4.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.11.7
python-dotenv==1.1.1
slowapi==0.1.9
pytest==8.4.1
httpx==0.28.1
alembic==1.16.4
